Switch button for FAB
