+++ lib/ui/group/members/debt_redirect/redirect_debt_view.dart
@@ -5,7 +5,7 @@ import 'package:statera/business_logic/debt_redirection/debt_redirection_cubit.d
 import 'package:statera/business_logic/layout/layout_state.dart';
 import 'package:statera/data/models/models.dart';
 import 'package:statera/ui/group/group_builder.dart';
-import 'package:statera/ui/group/members/debt_redirect/redirect_debt_commit_button.dart';
+import 'package:statera/ui/group/members/debt_redirect/redirect_debt_fab.dart';
 import 'package:statera/ui/group/members/debt_redirect/redirect_debt_header_text.dart';
 import 'package:statera/ui/group/members/debt_redirect/redirect_debt_visual.dart';
 import 'package:statera/ui/widgets/page_scaffold.dart';
@@ -20,20 +20,20 @@ class RedirectDebtView extends StatelessWidget {
     final isWide = context.select((LayoutState state) => state.isWide);
     final uid = context.read<AuthBloc>().uid;
 
-    return PageScaffold(
-      title: 'Redirect Debt',
-      child: Padding(
-        padding: EdgeInsets.symmetric(
-          horizontal: isWide ? MediaQuery.of(context).size.width / 3 : 30,
-        ),
-        child: GroupBuilder(
-          builder: (context, group) {
-            final debtRedirectionCubit = DebtRedirectionCubit(
-              uid: uid,
-              group: group,
-            );
-            return BlocProvider.value(
-              value: debtRedirectionCubit,
+    return GroupBuilder(
+      builder: (context, group) {
+        return BlocProvider.value(
+          value: DebtRedirectionCubit(
+            uid: uid,
+            group: group,
+          ),
+          child: PageScaffold(
+            title: 'Redirect Debt',
+            fab: RedirectDebtFAB(),
+            child: Padding(
+              padding: EdgeInsets.symmetric(
+                horizontal: isWide ? MediaQuery.of(context).size.width / 3 : 30,
+              ),
               child: Column(
                 mainAxisSize: MainAxisSize.min,
                 crossAxisAlignment: CrossAxisAlignment.stretch,
@@ -41,14 +41,12 @@ class RedirectDebtView extends StatelessWidget {
                   RedirectDebtHeaderText(),
                   RedirectDebtVisual(),
                   RedirectDebtVisual(isAfter: true),
-                  SizedBox(height: 50),
-                  RedirectDebtCommitButton(),
                 ],
               ),
-            );
-          },
-        ),
-      ),
+            ),
+          ),
+        );
+      },
     );
   }
 }
